Model de dificuldade de tarefa

+++ db/migrate/20160521131137_add_experience_to_task.rb
@@ -1,5 +1,5 @@
 class AddExperienceToTask < ActiveRecord::Migration
   def change
-    add_column :tasks, :experience, :integer
+    add_column :tasks, :experience, :decimal
   end
 end
