+++ db/migrate/20160521131430_add_experience_to_user.rb
@@ -1,5 +1,5 @@
 class AddExperienceToUser < ActiveRecord::Migration
   def change
-    add_column :users, :experience, :integer, default: 0
+    add_column :users, :experience, :decimal, default: 0
   end
 end
